(macOS) Improved error handling for the dependency build scripts.

+++ tools/macOS_dependencies_build.sh
@@ -34,6 +34,12 @@ echo "Building all dependencies in the ./external directory...\n"
 export PKG_CONFIG_PATH=$(pwd)/../local_install/lib/pkgconfig
 
 echo "Building libpng"
+
+if [ ! -d libpng ]; then
+  echo "libpng directory is missing, aborting."
+  exit
+fi
+
 cd libpng
 rm -f CMakeCache.txt
 if [ $(uname -m) == "arm64" ]; then
@@ -47,6 +53,12 @@ make install
 cd ..
 
 echo "\nBuilding FreeType"
+
+if [ ! -d freetype/build ]; then
+  echo "FreeType directory is missing, aborting."
+  exit
+fi
+
 cd freetype/build
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_DISABLE_FIND_PACKAGE_HarfBuzz=on -DBUILD_SHARED_LIBS=on -DCMAKE_MACOSX_RPATH=on -DCMAKE_INSTALL_PREFIX=$(pwd)/../../local_install -S .. -B .
@@ -56,6 +68,12 @@ cp libfreetype.6.18.0.dylib ../../../libfreetype.6.dylib
 cd ../..
 
 echo "\nBuilding FreeImage"
+
+if [ ! -d freeimage/FreeImage ]; then
+  echo "FreeImage directory is missing, aborting."
+  exit
+fi
+
 cd freeimage/FreeImage
 make clean
 make -j${JOBS}
@@ -63,6 +81,12 @@ cp libfreeimage.a ../../..
 cd ../..
 
 echo "\nBuilding pugixml"
+
+if [ ! -d pugixml ]; then
+  echo "pugixml directory is missing, aborting."
+  exit
+fi
+
 cd pugixml
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release .
@@ -72,6 +96,12 @@ cp libpugixml.a ../..
 cd ..
 
 echo "\nBuilding SDL"
+
+if [ ! -d SDL/build ]; then
+  echo "SDL directory is missing, aborting."
+  exit
+fi
+
 cd SDL/build
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release -S .. -B .
@@ -81,6 +111,12 @@ cp libSDL2-2.0.dylib ../../..
 cd ../..
 
 echo "\nBuilding libvpx"
+
+if [ ! -d libvpx ]; then
+  echo "libvpx directory is missing, aborting."
+  exit
+fi
+
 cd libvpx
 ./configure --disable-examples --disable-docs --enable-pic --enable-vp9-highbitdepth --prefix=$(pwd)/../local_install
 make clean
@@ -89,6 +125,12 @@ make install
 cd ..
 
 echo "\nBuilding Ogg"
+
+if [ ! -d ogg ]; then
+  echo "Ogg directory is missing, aborting."
+  exit
+fi
+
 cd ogg
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$(pwd)/../local_install .
@@ -98,6 +140,12 @@ make install
 cd ..
 
 echo "\nBuilding Vorbis"
+
+if [ ! -d vorbis ]; then
+  echo "Vorbis directory is missing, aborting."
+  exit
+fi
+
 cd vorbis
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release -DBUILD_SHARED_LIBS=on -DCMAKE_MACOSX_RPATH=on -DCMAKE_INSTALL_PREFIX=$(pwd)/../local_install .
@@ -109,6 +157,12 @@ cp lib/libvorbis.0.4.9.dylib ../..
 cd ..
 
 echo "\nBuilding Opus"
+
+if [ ! -d opus ]; then
+  echo "Opus directory is missing, aborting."
+  exit
+fi
+
 cd opus
 rm -f CMakeCache.txt
 cmake -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$(pwd)/../local_install .
@@ -118,6 +172,12 @@ make install
 cd ..
 
 echo "\nBuilding FFmpeg"
+
+if [ ! -d FFmpeg ]; then
+  echo "FFmpeg directory is missing, aborting."
+  exit
+fi
+
 cd FFmpeg
 PKG_CONFIG_PATH=$(pwd)/../local_install/lib/pkgconfig ./configure --prefix=/usr/local --enable-rpath --install-name-dir=@rpath --disable-doc --enable-gpl --enable-shared --enable-libvorbis --enable-libopus --enable-libvpx --enable-postproc
 
@@ -139,3 +199,6 @@ install_name_tool -rpath /usr/local/lib @executable_path libswscale/libswscale.6
 cp libswscale/libswscale.6.dylib ../..
 
 unset PKG_CONFIG_PATH
+
+echo
+echo "Done building all dependencies."
